fix: removeScrapedAnime 로직 변경
animeId가 아닌 scrapId로 조회하도록 변경

+++ src/modules/scrap/scrap.service.ts
@@ -1,12 +1,15 @@
 import {
   ForbiddenException,
   Injectable,
-  UnauthorizedException,
+  NotFoundException,
 } from '@nestjs/common';
 import { User } from '../user/entities/user.entity';
 import { ResponseDto } from '../../common/dto/response.dto';
 import { EStatusCode } from '../../common/enum/status.enum';
-import { EResponseMessage } from '../../common/enum/message.enum';
+import {
+  EErrorMessage,
+  EResponseMessage,
+} from '../../common/enum/message.enum';
 import { ScrapRepository } from './repository/scrap.repository';
 import { AnimeRepository } from '../anime/repository/anime.repository';
 
@@ -39,13 +42,15 @@ export class ScrapService {
     );
   }
 
-  async removeScrapedAnime(animeId: number, user: User) {
-    const scrap = await this.scrapRepository.getScrapsByIds(animeId, user.id, [
-      'user',
-    ]);
+  async removeScrapedAnime(scrapId: number, user: User) {
+    const scrap = await this.scrapRepository.getScrapById(scrapId, ['user']);
+
+    if (!scrap) {
+      throw new NotFoundException(EErrorMessage.NOT_FOUND);
+    }
 
     if (scrap.user.id !== user.id) {
-      return new UnauthorizedException('누구십니까');
+      throw new ForbiddenException(EErrorMessage.NOT_PERMISSIONS);
     }
 
     await this.scrapRepository.remove(scrap);
